backend-ai/requirements.txt
flask
flask-cors
gunicorn
python-dotenv
 langchain
langchain-groq
